@@ -1,9 +1,6 @@
 package com.wbjackson.youtubeweatherapp;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -14,6 +11,9 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.wbjackson.youtubeweatherapp.CurrentWeatherService.CurrentWeatherCallback;
 
 public class WeatherActivity extends AppCompatActivity {
